reorganizes runtime files
