+++ dlt/common/__init__.py
@@ -2,4 +2,5 @@ from dlt.common.arithmetics import Decimal  # noqa: F401
 from dlt.common.wei import Wei
 from dlt.common.pendulum import pendulum  # noqa: F401
 from dlt.common.json import json  # noqa: F401, I251
-from dlt.common.signals import sleep   # noqa: F401
+from dlt.common.runtime.signals import sleep   # noqa: F401
+from dlt.common.runtime import logger  # noqa: F401
