+++ dlt/common/configuration/specs/__init__.py
@@ -3,7 +3,6 @@ from .base_configuration import BaseConfiguration, CredentialsConfiguration, Cre
 from .normalize_volume_configuration import NormalizeVolumeConfiguration  # noqa: F401
 from .load_volume_configuration import LoadVolumeConfiguration  # noqa: F401
 from .schema_volume_configuration import SchemaVolumeConfiguration, TSchemaFileFormat  # noqa: F401
-from .pool_runner_configuration import PoolRunnerConfiguration, TPoolType  # noqa: F401
 from .gcp_client_credentials import GcpClientCredentials, GcpClientCredentialsWithDefault  # noqa: F401
 from .postgres_credentials import PostgresCredentials, RedshiftCredentials, ConnectionStringCredentials  # noqa: F401
 from .config_section_context import ConfigSectionContext  # noqa: F401
