+++ dlt/load/load.py
@@ -7,12 +7,12 @@ from dlt.common.configuration import with_config, known_sections
 from dlt.common.configuration.accessors import config
 from dlt.common.typing import StrAny
 from dlt.common.runners import TRunMetrics, Runnable, workermethod
-from dlt.common.logger import pretty_format_exception
+from dlt.common.runtime.logger import pretty_format_exception
 from dlt.common.exceptions import TerminalValueError
 from dlt.common.schema import Schema
 from dlt.common.schema.typing import TTableSchema
 from dlt.common.storages import LoadStorage
-from dlt.common.telemetry import get_logging_extras, set_gauge_all_labels
+from dlt.common.runtime.telemetry import get_logging_extras, set_gauge_all_labels
 from dlt.common.destination.reference import JobClientBase, DestinationReference, LoadJob, TLoadJobStatus, DestinationClientConfiguration
 
 from dlt.destinations.job_client_impl import LoadEmptyJob
