+++ dlt/normalize/normalize.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, List, Dict, Sequence, Tuple
 from multiprocessing.pool import AsyncResult, Pool as ProcessPool
 from prometheus_client import Counter, CollectorRegistry, REGISTRY, Gauge
 
-from dlt.common import pendulum, signals, json, logger, sleep
+from dlt.common import pendulum, json, logger, sleep
 from dlt.common.configuration import with_config, known_sections
 from dlt.common.configuration.accessors import config
 from dlt.common.configuration.container import Container
@@ -12,10 +12,11 @@ from dlt.common.configuration.specs import LoadVolumeConfiguration, NormalizeVol
 from dlt.common.destination import DestinationCapabilitiesContext, TLoaderFileFormat
 from dlt.common.json import custom_pua_decode
 from dlt.common.runners import TRunMetrics, Runnable
+from dlt.common.runtime import signals
 from dlt.common.schema.typing import TStoredSchema, TTableSchemaColumns
 from dlt.common.storages.exceptions import SchemaNotFoundError
 from dlt.common.storages import NormalizeStorage, SchemaStorage, LoadStorage
-from dlt.common.telemetry import get_logging_extras
+from dlt.common.runtime.telemetry import get_logging_extras
 from dlt.common.typing import TDataItem
 from dlt.common.schema import TSchemaUpdate, Schema
 from dlt.common.schema.exceptions import CannotCoerceColumnException
