+++ tests/common/runners/utils.py
@@ -7,7 +7,7 @@ from multiprocessing.pool import Pool
 
 from dlt.common import logger
 from dlt.common.runners.runnable import Runnable, workermethod
-from dlt.common.telemetry import TRunMetrics
+from dlt.common.runtime.telemetry import TRunMetrics
 from dlt.common.utils import uniq_id
 
 ALL_METHODS = multiprocessing.get_all_start_methods()
