+++ tests/common/scripts/stdout_encode_result.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 from dlt.common.runners.stdout import exec_to_stdout
-from dlt.common.telemetry import TRunMetrics
+from dlt.common.runtime.telemetry import TRunMetrics
 
 
 def worker(data1, data2):
